1.添加register线程测试，thrift连通。 2.修复了updateNodeMap中一处应return的错误，以及一处字符串分割错误 3.添加request线程测试（getsuitable），thrift连通 4.添加requestServiceSituation测试，thrift连通 5.更改computerStatus数据字段为String，便于传输，在server端的get方法中将String转化为相应的变量
现已实现：
1.SC启动后，若有client通过thrift调用registerService，SC将添加对该Service及其Path的Watch，并对Watch返回的list中所有节点获取机器信息（内存，CPU，磁盘用量）定时更新
2.SC启动后，若有client通过thrift调用getSuitable，thrift将返回相应服务组中内存使用率最小（目前）的节点。

下一版本对Ruler模块进行细化完善

+++ src/main/java/mobi/bihu/crawler/sc/loadbalance/Ruler.java
@@ -26,11 +26,11 @@ class Ruler {
     }
 
     String findSuitable(ConcurrentHashMap<Item,ComputerStatus> list){
-        int min = Integer.MAX_VALUE;
+        double min = Double.MAX_VALUE;
         String path = null;
         Item minItem = null;
         for (Map.Entry<Item, ComputerStatus> entry : list.entrySet()) {
-            int value = entry.getValue().getMemory();
+            double value = entry.getValue().getMemory();
             if(value < min){
                 min = value;
                 minItem = entry.getKey();
